pyproject.toml
[project]
name = "webserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "e2b-code-interpreter>=2.3.0",
    "openai>=2.8.1",
]
